Work on node and links selection & deletion

+++ NodeEye/src/main/kotlin/io/github/deltacv/nodeeye/node/Node.kt
@@ -4,31 +4,43 @@ import imgui.ImGui
 import io.github.deltacv.nodeeye.id.DrawableIdElement
 import io.github.deltacv.nodeeye.id.IdElementContainer
 import io.github.deltacv.nodeeye.attribute.Attribute
+import io.github.deltacv.nodeeye.attribute.AttributeMode
 
-abstract class Node : DrawableIdElement {
+abstract class Node(protected var allowDelete: Boolean = true) : DrawableIdElement {
 
     companion object {
         val nodes = IdElementContainer<Node>()
         val attributes = IdElementContainer<Attribute>()
+
+        @JvmStatic protected val INPUT = AttributeMode.INPUT
+        @JvmStatic protected val OUTPUT = AttributeMode.OUTPUT
     }
 
     override val id by nodes.nextId { this }
 
     val nodeAttributes = mutableListOf<Attribute>()
 
-    fun drawAttributes() {
-        for(attribute in nodeAttributes) {
+    protected fun drawAttributes() {
+        for((i, attribute) in nodeAttributes.withIndex()) {
+            attribute.parentNode = this
             attribute.draw()
-            //ImGui.separator() TODO separate
+
+            if(i < nodeAttributes.size - 1) {
+                ImGui.newLine() // make a new blank line if this isn't the last attribute
+            }
         }
     }
 
     override fun delete() {
-        for(attribute in nodeAttributes) {
-            attribute.delete()
-        }
+        if(allowDelete) {
+            for (attribute in nodeAttributes) {
+                attribute.delete()
+            }
 
-        nodes.removeId(id)
+            nodes.removeId(id)
+        }
     }
+
+    operator fun Attribute.unaryPlus() = nodeAttributes.add(this)
 
 }
